src/requirements.txt
boto3==1.29.3
Pillow==10.1.0
requests==2.31.0

